refactor: 👷 remove ProjectPageLayout
