+++ components/layout/projects/ProjectPageLayoutWrapper.tsx
@@ -1,20 +1,29 @@
 import { FC, ReactNode } from 'react'
 
-import ProjectPageLayout from '@/components/layout/projects/ProjectPageLayout'
+import PageContainer from '@/components/layout/PageContainer'
+import Gallery from '@/components/layout/projects/project-page/Gallery'
+import HeaderSection from '@/components/layout/projects/project-page/HeaderSection'
+import ProjectInformation from '@/components/layout/projects/project-page/ProjectInformation'
+import ProjectResponsibility from '@/components/layout/projects/project-page/ProjectResponsibility'
+import BreadCrumbs from '@/components/shared/Breadcrumbs'
+import CallToActionResume from '@/components/shared/call-to-action/CallToActionResume'
 
-import { GoBackLinkEnum } from '@/lib/utils/interfaces/enums'
+import { HeaderSectionProps } from '@/lib/utils/interfaces/componentProps'
 import { Project, Section } from '@/lib/utils/interfaces/interfaces'
 import { BreadCrumbsType } from '@/lib/utils/interfaces/types'
 
-type ProjectPageLayoutWrapperProps = {
+type BreadCrumbsProps = {
   breadCrumbs: BreadCrumbsType
+}
+
+type PageProps = {
   pageID: string
-  goBackLink: GoBackLinkEnum
-  sectionID: string
   projectData: Project
   sections: Section[]
   PageNavigation: ReactNode
 }
+
+type ProjectPageLayoutWrapperProps = BreadCrumbsProps & HeaderSectionProps & PageProps
 
 const ProjectPageLayoutWrapper: FC<ProjectPageLayoutWrapperProps> = ({
   breadCrumbs,
@@ -25,6 +34,8 @@ const ProjectPageLayoutWrapper: FC<ProjectPageLayoutWrapperProps> = ({
   sections,
   PageNavigation,
 }): JSX.Element => {
+  const { level1Url, level1Text, level2Url, level2Text } = breadCrumbs
+
   const {
     title,
     role,
@@ -39,24 +50,34 @@ const ProjectPageLayoutWrapper: FC<ProjectPageLayoutWrapperProps> = ({
   } = projectData
 
   return (
-    <ProjectPageLayout
-      breadCrumbs={breadCrumbs}
-      pageID={pageID}
-      title={title}
-      role={role}
-      years={years}
-      company={company}
-      goBackLink={goBackLink}
-      sectionID={sectionID}
-      description={description}
-      skillsOverview={skillsOverview}
-      customers={customers}
-      projectLinks={projectLinks}
-      linkGitHub={linkGitHub}
-      sections={sections}
-      imageShowcase={imageShowcase}
-      PageNavigation={PageNavigation}
-    />
+    <>
+      <PageContainer id={pageID}>
+        <BreadCrumbs level1Url={level1Url} level1Text={level1Text} level2Url={level2Url} level2Text={level2Text} />
+        <HeaderSection
+          title={title}
+          role={role}
+          years={years}
+          company={company}
+          goBackLink={goBackLink}
+          sectionID={sectionID}
+        />
+        <article>
+          <ProjectInformation
+            description={description}
+            skillsOverview={skillsOverview}
+            customers={customers}
+            projectLinks={projectLinks}
+            linkGitHub={linkGitHub}
+          />
+          <ProjectResponsibility sections={sections} />
+          <Gallery imageShowcase={imageShowcase} />
+
+          <CallToActionResume />
+        </article>
+
+        {PageNavigation}
+      </PageContainer>
+    </>
   )
 }
