+++ lib/utils/interfaces/componentProps.ts
@@ -1,9 +1,9 @@
 import { GoBackLinkEnum } from '@/lib/utils/interfaces/enums'
 
 export type HeaderSectionProps = {
-  title: string
+  title?: string
   sectionID: string
-  role: string
+  role?: string
   years?: string | undefined
   company?: string
   goBackLink: GoBackLinkEnum
